migraciones

+++ database/migrations/2021_01_24_175627_create_users_table.php
@@ -19,7 +19,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->integer('idSucursal')->unsigned();
+            $table->bigInteger('idSucursal')->unsigned();
             $table->foreign('idSucursal')->references('id')->on('sucursales');
             $table->boolean('condicion')->default(1);
             $table->rememberToken();
