+++ database/migrations/2021_01_24_180121_create_stocks_table.php
@@ -16,9 +16,9 @@ class CreateStocksTable extends Migration
         Schema::create('stocks', function (Blueprint $table) {
             $table->id();
             $table->integer('existencia');
-            $table->integer('idSucursal')->unsigned();
+            $table->bigInteger('idSucursal')->unsigned();
             $table->foreign('idSucursal')->references('id')->on('sucursales');
-            $table->integer('idProducto')->unsigned();
+            $table->bigInteger('idProducto')->unsigned();
             $table->foreign('idProducto')->references('id')->on('productos');
             $table->boolean('condicion')->default(1);
             $table->timestamps();
