+++ database/migrations/2021_01_24_180159_create_vencidos_table.php
@@ -16,9 +16,9 @@ class CreateVencidosTable extends Migration
         Schema::create('vencidos', function (Blueprint $table) {
             $table->id();
             $table->integer('caducado');
-            $table->integer('idSucursal')->unsigned();
+            $table->bigInteger('idSucursal')->unsigned();
             $table->foreign('idSucursal')->references('id')->on('sucursales');
-            $table->integer('idProducto')->unsigned();
+            $table->bigInteger('idProducto')->unsigned();
             $table->foreign('idProducto')->references('id')->on('productos');
             $table->boolean('condicion')->default(1);
             $table->timestamps();
