+++ database/migrations/2021_01_24_180231_create_compras_table.php
@@ -15,9 +15,9 @@ class CreateComprasTable extends Migration
     {
         Schema::create('compras', function (Blueprint $table) {
             $table->id();
-            $table->integer('idProveedor')->unsigned();
+            $table->bigInteger('idProveedor')->unsigned();
             $table->foreign('idProveedor')->references('id')->on('proveedores');
-            $table->integer('idSucursal')->unsigned();
+            $table->bigInteger('idSucursal')->unsigned();
             $table->foreign('idSucursal')->references('id')->on('sucursales');
             $table->string('numcompra',15)->nullable();
             $table->dateTime('fechacom');
